add ROWS_OF_RECORDS_BY_PAGE constant.
