+++ src/api/ADempiere/user-interface/persistence.js
@@ -17,8 +17,9 @@
 // Get Instance for connection
 import { request } from '@/utils/ADempiere/request'
 import { isEmptyValue } from '@/utils/ADempiere/valueUtils'
+
 // constants
-import { ROW_ATTRIBUTES } from '@/utils/ADempiere/constants/table'
+import { ROWS_OF_RECORDS_BY_PAGE } from '@/utils/ADempiere/constants/table'
 
 /**
  * Object List from window
@@ -39,7 +40,7 @@ export function getEntities({
   searchValue = '',
   filters = [],
   pageToken,
-  pageSize
+  pageSize = ROWS_OF_RECORDS_BY_PAGE
 }) {
   // used as where clause
   if (!isEmptyValue(filters)) {
@@ -91,7 +92,7 @@ export function getEntities({
       sorting: sortingDefinition,
       // Page Data
       page_token: pageToken,
-      page_size: ROW_ATTRIBUTES.pageSize
+      page_size: pageSize
     }
   })
     .then(response => {
