+++ src/store/modules/ADempiere/pointOfSales/productPrice/actions.js
@@ -14,10 +14,14 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+// api request methods
 import {
   getProductPriceList
 } from '@/api/ADempiere/form/point-of-sales.js'
-import { isEmptyValue, extractPagingToken } from '@/utils/ADempiere/valueUtils.js'
+
+// utils and helper methods
+import { isEmptyValue } from '@/utils/ADempiere/valueUtils.js'
+import { extractPagingToken } from '@/utils/ADempiere/dataUtils'
 import { showMessage } from '@/utils/ADempiere/notification.js'
 import language from '@/lang'
