+++ src/utils/ADempiere/valueUtils.js
@@ -146,37 +146,6 @@ export function castValueWithType({
 }
 
 /**
- * Return token of pagination.
- * @author EdwinBetanc0urt <EdwinBetanc0urt@oulook.com>
- * @param {string} token
- * @returns {string}
- */
-export function extractPagingToken(token) {
-  if (isEmptyValue(token)) {
-    return ''
-  }
-
-  let onlyToken = token.slice(0, -2)
-  if (onlyToken.substr(-1, 1) === '-') {
-    // removes end hyphen
-    onlyToken = onlyToken.slice(0, -1)
-  }
-  return onlyToken
-}
-
-export function generatePageToken({ pageNumber = 1, token }) {
-  if (pageNumber < 1) {
-    pageNumber = 1
-  }
-
-  const onlyToken = extractPagingToken(token)
-  if (isEmptyValue(onlyToken)) {
-    return ''
-  }
-
-  return onlyToken + '-' + pageNumber
-}
-/**
  * @param {number} number
  */
 export function round(number, standardPrecision) {
